refactor(tools): break tools/index <-> tools/* cycle via deps.ts
The per-tool files imported `FreelanceToolDeps` from `./index.js`,
and `./index.js` imported each `register*Tool` function from the
per-tool files — a module cycle. TypeScript resolves it at build
time because the import on the tool-file side is `import type` only
(no runtime evaluation), but cycles are still a code smell, they
trip some bundlers, and they're exactly what static-analysis tools
like madge/knip flag.

Extract `FreelanceToolDeps` into `src/tools/deps.ts`, a leaf module
that imports only from ../sources, ../state, ../types. Every
per-tool file now imports the type from `./deps.js`. `tools/index.ts`
imports from `./deps.js` itself and re-exports the type for external
consumers so the public surface of `./tools/index.js` is unchanged.

Dependency graph after the change:

    deps.ts  ←  (every tool file)   ←  index.ts
                                      index.ts also imports deps.ts

No cycles: deps.ts depends on nothing inside ./tools/, tool files
depend only on deps.ts, index.ts depends on deps.ts + every tool
file with no back-edges.

11 files touched in src/tools/ (10 tool files with the import path
rewritten, 1 new deps.ts, 1 edit to index.ts that drops the inline
interface and re-exports it). No behavior change. 569/569 tests
pass, build clean, biome clean.
