+++ src/tools/advance.ts
@@ -1,7 +1,7 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { z } from "zod";
 import { errorResponse, handleError, jsonResponse } from "../mcp-helpers.js";
-import type { FreelanceToolDeps } from "./index.js";
+import type { FreelanceToolDeps } from "./deps.js";
 
 export function registerAdvanceTool(server: McpServer, deps: FreelanceToolDeps): void {
   const { manager } = deps;
