+++ src/tools/list.ts
@@ -1,6 +1,6 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { handleError, jsonResponse } from "../mcp-helpers.js";
-import type { FreelanceToolDeps } from "./index.js";
+import type { FreelanceToolDeps } from "./deps.js";
 
 export function registerListTool(server: McpServer, deps: FreelanceToolDeps): void {
   const { manager, getLoadErrors } = deps;
