+++ src/tools/start.ts
@@ -2,7 +2,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { z } from "zod";
 import { handleError, jsonResponse } from "../mcp-helpers.js";
 import { validateGraphSources } from "../sources.js";
-import type { FreelanceToolDeps } from "./index.js";
+import type { FreelanceToolDeps } from "./deps.js";
 
 export function registerStartTool(server: McpServer, deps: FreelanceToolDeps): void {
   const { manager, graphs, sourceOpts, validateSourcesOnStart } = deps;
